+++ src/searchForm/datePicker/section/DatePickerSection.tsx
@@ -1,30 +1,38 @@
-import React from 'react'
+import React, { useRef } from 'react'
 import Datepicker, { DatePickerProps, DatePickerOrientation } from 'datePicker'
 import Divider from 'divider'
 import Section from 'layout/section/baseSection'
 import ChevronIcon from 'icon/chevronIcon'
 import Item from '_utils/item'
+import { useFocusTrap } from '_utils/useFocusTrap'
 
 export interface DatePickerSectionProps extends DatePickerProps {
-  readonly title: string
-  readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void
+  title: string
+  onClose: () => void
 }
 
-export const DatePickerSection = (props: DatePickerSectionProps) => (
-  <Section className={props.className}>
-    <Item
-      leftAddon={<ChevronIcon left />}
-      leftTitle={props.title}
-      tag={<button type="button" />}
-      onClick={props.onClick}
-    />
-    <Divider />
-    <Datepicker
-      {...props}
-      className="kirk-datePickerSection-datepicker"
-      orientation={DatePickerOrientation.VERTICAL}
-    />
-  </Section>
-)
+export const DatePickerSection = ({ onClose, ...props }: DatePickerSectionProps) => {
+  const ref = useRef<HTMLDivElement>(null)
+  useFocusTrap(ref, onClose)
+
+  return (
+    <div ref={ref} role="dialog" className={props.className}>
+      <Section>
+        <Item
+          leftAddon={<ChevronIcon left />}
+          leftTitle={props.title}
+          tag={<button type="button" />}
+          onClick={onClose}
+        />
+        <Divider />
+        <Datepicker
+          {...props}
+          className="kirk-datePickerSection-datepicker"
+          orientation={DatePickerOrientation.VERTICAL}
+        />
+      </Section>
+    </div>
+  )
+}
 
 export default DatePickerSection